Renamed modules. Added dummy implementation for new interface functions
