+++ consensus/hotstuff/safetyrules/voter.go
@@ -1,10 +1,11 @@
-package voter
+package safetyrules
 
 import (
 	"fmt"
 
 	"github.com/onflow/flow-go/consensus/hotstuff"
 	"github.com/onflow/flow-go/consensus/hotstuff/model"
+	"github.com/onflow/flow-go/model/flow"
 )
 
 // Voter produces votes for the given block
@@ -83,4 +84,25 @@ func (v *Voter) ProduceVote(proposal *model.Proposal, curView uint64) (*model.Vo
 	}
 
 	return vote, nil
+}
+
+// ProduceTimeout takes current view, highest locally known QC and TC and decides whether to produce timeout for current view.
+// Returns:
+//  * (timeout, nil): On the _first_ block for the current view that is safe to vote for.
+//    Subsequently, voter does _not_ vote for any other block with the same (or lower) view.
+//  * (nil, model.NoTimeoutError): If the safety module decides that it is not safe to timeout under current conditions.
+//    This is a sentinel error and _expected_ during normal operation.
+// All other errors are unexpected and potential symptoms of uncovered edge cases or corrupted internal state (fatal).
+func (v *Voter) ProduceTimeout(curView uint64, highestQC *flow.QuorumCertificate, highestTC *flow.TimeoutCertificate) (*model.TimeoutObject, error) {
+	panic("to be implemented")
+}
+
+// IsSafeToVote checks if this proposal is valid in terms of voting rules, if voting for this proposal won't break safety rules.
+func (v *Voter) IsSafeToVote(proposal *model.Proposal) bool {
+	panic("to be implemented")
+}
+
+// IsSafeToTimeout checks if it's safe to timeout with proposed data, if timing out won't break safety rules.
+func (v *Voter) IsSafeToTimeout(curView uint64, highestQC *flow.QuorumCertificate, highestTC *flow.TimeoutCertificate) bool {
+	panic("to be implemented")
 }
