+++ consensus/hotstuff/safetyrules/voter_test.go
@@ -1,4 +1,4 @@
-package voter
+package safetyrules
 
 import (
 	"testing"
